+++ python/rlgym_learn_algos/ppo/discrete_actor.py
@@ -24,11 +24,11 @@ class DiscreteFF(Actor[AgentID, np.ndarray, np.ndarray]):
         self,
         input_size: int,
         n_actions: int,
-        layer_sizes: tuple[int],
-        device: torch.Device,
+        layer_sizes: tuple[int, ...],
+        device: torch.device,
     ):
         super().__init__()
-        self.device: torch.Device = device
+        self.device: torch.device = device
 
         assert len(layer_sizes) != 0, (
             "AT LEAST ONE LAYER MUST BE SPECIFIED TO BUILD THE NEURAL NETWORK!"